Updated Frontend with React and Added Minimal Sentiment Analysis
I had to remake the whole frontend module and translate it to react, as I figured this would structure the project better. Now everything works in a React frontend. I also started working on some starting sentiment analysis with pre-trained hugging face transformers. I also fixed the search functionality and added a backend that receives a product search query for the system to process.