Added backend and Got System to correctly serve results
I have made some adjustments to the React frontend so it functions better and more smoothly. I have also linked up a backend for the React frontend, so when a user enters a search, it is sent to the backend, processed by all of the scripts and then the API sends a repsonse back to be displayed in the frontend. The user flow works.